feat: Change split char for multi-element fields
Originally, the library dataset used a slash to indicate that a field contained multiple data points, as in `"Address": "632 S Columbus/813 Gibbon"`. This approach is problematic for parsing, as there are also street addresses such as `918 1/2 Pendleton` that need to use a slash to show the address. Fields containing multiple data points now seperate the data using a comma, like ` "Address": "521 N Royal, 417 N Fayette"`